requirements.txt
google-cloud-speech
google-api-python-client
pydub
librosa
soundfile